requirements.txt
asgiref==3.8.1
blinker==1.9.0
bottle==0.13.2
click==8.1.8
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
proxy_tools==0.1.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
PyQtWebEngine==5.15.7
PyQtWebEngine-Qt5==5.15.16
PySide2==5.13.2
pywebview==5.3.2
QtPy==2.4.2
shiboken2==5.13.2
typing_extensions==4.12.2
Werkzeug==3.1.3
